Move verifyRequest method into util trait
So it can be used for other paths too
